requirements.txt
django 
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-filters